requirements.txt
flask==3.0.2
pyodbc==5.1.0
python-dotenv==1.0.1
openpyxl==3.1.5
python-docx==1.1.2
matplotlib==3.8.4
